feat: Add quest functionality and Unsplash API integration
- Implement quest.js for quest handling
- Add Unsplash API integration for image fetching
- Update index and newQuest HTML templates
- Modify main.js to support new features
