+++ src/html/index.html
@@ -28,5 +28,6 @@
       </div>
     </main>
     <script src="../js/main.js" type="module"></script>
+    <script src="../js/quest.js" type="module"></script>
   </body>
 </html>
