+++ src/html/newQuest.html
@@ -84,5 +84,6 @@
     </div>
 
     <script src="../js/main.js" type="module"></script>
+    <script src="../js/quest.js" type="module"></script>
   </body>
 </html>
